fix: data corrections and test fixes
- Correct typos in effects and affixes data (Poison, damage spelling)
- Fix test property names and missing StateManager registrations
- Add .gitignore entry for node modules
- Improve test coverage for critical hit mechanics
- Update changelog with P1S3 completion details

+++ src/engine/hit_context.py
@@ -19,7 +19,7 @@ class HitContext:
     # Core damage values
     base_raw: Any        # Original damage input (int, tuple, etc.)
     base_resolved: int   # Final resolved base damage value
-    final_damage: int    # Final damage after all calculations
+    final_damage: float    # Final damage after all calculations
 
     # Combat outcome flags (Phase 2 mechanics)
     was_crit: bool = False
